+++ source/DasBlog.SmokeTest/Browser.cs
@@ -1,0 +1,8 @@
+﻿using DasBlog.SmokeTest.Interfaces;
+
+namespace DasBlog.SmokeTest
+{
+	internal class Browser : IBrowser
+	{
+	}
+}
