+++ source/DasBlog.SmokeTest/DasBlogInstallationOptions.cs
@@ -1,0 +1,7 @@
+﻿namespace DasBlog.SmokeTest
+{
+	internal class DasBlogInstallationOptions
+	{
+		public string ContentRootPath { get; set; }
+	}
+}
