+++ source/DasBlog.SmokeTest/GitVersionedFileServiceOptions.cs
@@ -1,0 +1,7 @@
+﻿namespace DasBlog.SmokeTest
+{
+	internal class GitVersionedFileServiceOptions
+	{
+		public string GitRepo { get; set; }
+	}
+}
