+++ source/DasBlog.SmokeTest/Interfaces/IBrowser.cs
@@ -1,0 +1,6 @@
+﻿namespace DasBlog.SmokeTest.Interfaces
+{
+	public interface IBrowser
+	{
+	}
+}
