+++ source/DasBlog.SmokeTest/Interfaces/IDasBlogInstallation.cs
@@ -1,0 +1,12 @@
+﻿namespace DasBlog.SmokeTest.Interfaces
+{
+	public interface IDasBlogInstallation
+	{
+		void Init();
+		void Terminate();
+		string GetConfigPathAndFile();
+		string GetContentDirectoryPath();
+		string GetLogDirectoryPath();
+		string GetWwwRootDirectoryPath();
+	}
+}
