+++ source/DasBlog.SmokeTest/Interfaces/ITester.cs
@@ -1,0 +1,7 @@
+﻿namespace DasBlog.SmokeTest.Interfaces
+{
+	public interface ITester
+	{
+		void Test();
+	}
+}
